Stop using BOSH InstanceGroup Properties
BOSH InstanceGroup Properties have been removed and are replaced by Job
Properties. This has broken CI testing.

The implementation and tests already relied on Redis being the first
instance group. They now also rely on Redis being the first job. That's
not a generally safe assumption, but for this adapter at this time is
happens to always be true.

[#168871114]